partial phasing out of getcells

+++ source/newtonian/one_dimensional/eulerian1d.cpp
@@ -1,7 +1,7 @@
 #include "eulerian1d.hpp"
 
 double Eulerian1D::operator()
-(int /*i*/, vector<double> const& /*vp*/,
+(size_t /*i*/, vector<double> const& /*vp*/,
  vector<ComputationalCell> const& /*hv*/) const
 {
   return 0;
