+++ source/newtonian/one_dimensional/eulerian1d.hpp
@@ -15,7 +15,7 @@ class Eulerian1D: public VertexMotion
 {
 public:
   
-  double operator()(int i, vector<double> const& vp,
+  double operator()(size_t i, vector<double> const& vp,
 		    vector<ComputationalCell> const& hv) const;
 };
